refactor: move operations to ESM
