+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+# v1.0.0
+
+- Moved to JavaScript modules
