+++ manifest.ttl
@@ -8,5 +8,5 @@
         rdfs:comment "Loads a CSVW file from the local filesystem or the web depending on the given argument" ;
         code:implementedBy [
                                a code:EcmaScript ;
-                               code:link <node:barnard59-csvw#fetch>
+                               code:link <node:barnard59-csvw/fetch.js>
                            ] .
